requirements.txt
flask
openai
python-dotenv
numpy
pypdf
python-docx